[B] Adjust various element transforms
