+++ client/src/global/components/form/ContentEditor/plugins/withHtmlAwareBreaks.js
@@ -1,6 +1,6 @@
 import { Editor, Range, Transforms, Element, Point, Node } from "slate";
-import { rteElements } from "../utils/elements";
-import { decreaseIndent } from "../utils/slate/listIndents";
+import { rteElements, nestableElements } from "../utils/elements";
+import { decreaseIndent } from "../utils/slate/transforms/listIndents";
 import { getListItemNode } from "../utils/slate/getters";
 
 // The plugins here provide sane defaults for dealing with enter/delete given that we have nested html structure, in particular nested blocks, which isn't accounted for in the default slate implementation.
@@ -29,7 +29,11 @@ const withHtmlAwareBreaks = editor => {
     // Remove all attributes other than the type/tag and marks, so we don't copy id, classes, anything that can't be updated in the RTE
     const { children, htmlAttrs, slateOnly, inline, ...next } = block;
     // Insert a paragraph if this node type isn't editable in the RTE; handle unwrapping the list item if we're on an empty li
-    if (!rteElements.includes(next.type)) next.type = "p";
+    if (
+      !rteElements.includes(next.type) &&
+      !nestableElements.includes(next.type)
+    )
+      next.type = "p";
     if (next.type === "a" || next.type === "img" || next.type === "iframe")
       next.type = "p";
     if (next.type === "li") {
