+++ client/src/global/components/form/ContentEditor/plugins/withShortcuts.js
@@ -1,5 +1,5 @@
 import { Editor, Element, Range, Transforms, Node } from "slate";
-import { toggleOrWrapBlock } from "../utils/slate/transforms";
+import { toggleOrWrapNode } from "../utils/slate/transforms";
 
 const SHORTCUTS = {
   "*": "ul",
@@ -49,7 +49,7 @@ const withShortcuts = editor => {
         if (!Range.isCollapsed(range)) {
           Transforms.delete(editor);
         }
-        return toggleOrWrapBlock(editor, blockType);
+        return toggleOrWrapNode(editor, blockType);
       }
     }
