+++ client/src/global/components/form/ContentEditor/utils/slate/transforms/index.js
@@ -1,0 +1,7 @@
+export { unwrapLink, wrapLink } from "./links";
+export { insertImage } from "./images";
+export { insertIframe } from "./iframes";
+export { toggleTextBlock } from "./textBlocks";
+export { insertOrWrapSpan, unwrapSpan } from "./spans";
+export { wrapLayoutBlock } from "./layoutBlocks";
+export { toggleOrWrapNode, unwrapNode, toggleMark } from "./shared";
