+++ client/src/global/components/form/ContentEditor/utils/slate/transforms/links.js
@@ -1,0 +1,29 @@
+import { Transforms, Editor, Range, Element } from "slate";
+import { ReactEditor } from "slate-react";
+
+export const unwrapLink = editor => {
+  Transforms.unwrapNodes(editor, {
+    match: n => !Editor.isEditor(n) && Element.isElement(n) && n.type === "a"
+  });
+  ReactEditor.focus(editor);
+};
+
+export const wrapLink = (editor, url, text) => {
+  const { selection } = editor ?? {};
+  if (!selection) return;
+  const isCollapsed = selection && Range.isCollapsed(selection);
+  const link = {
+    type: "a",
+    htmlAttrs: { href: url },
+    children: isCollapsed ? [{ text: text ?? url }] : []
+  };
+
+  if (isCollapsed) {
+    Transforms.insertNodes(editor, link);
+  } else {
+    Transforms.wrapNodes(editor, link, { split: true });
+    Transforms.collapse(editor, { edge: "end" });
+  }
+
+  ReactEditor.focus(editor);
+};
