+++ client/src/global/components/form/ContentEditor/utils/slate/transforms/removeNode.js
@@ -2,7 +2,7 @@ import { Transforms, Editor, Node, Element } from "slate";
 import { ReactEditor } from "slate-react";
 import has from "lodash/has";
 import isEqual from "lodash/isEqual";
-import { isEmptyAndChildless } from "./getters";
+import { isEmptyAndChildless } from "../getters";
 
 // This function removes a specific node at any point in the selection hierarchy, not just the lowest one like the other transforms. It is used to handle delete via the html element labels.
